# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuizBlast API - Kahoot Clone
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Quiz {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  createdAt: string;
  updatedAt: string;
  questionCount: number;
}

export interface CreateQuizBody {
  title: string;
  /** @nullable */
  description?: string | null;
}

export interface UpdateQuizBody {
  title?: string;
  /** @nullable */
  description?: string | null;
}

export interface Question {
  id: number;
  quizId: number;
  text: string;
  options: string[];
  correctOption: number;
  timeLimit: number;
  points: number;
  orderIndex: number;
}

export interface CreateQuestionBody {
  text: string;
  options: string[];
  correctOption: number;
  timeLimit?: number;
  points?: number;
  orderIndex?: number;
}

export interface UpdateQuestionBody {
  text?: string;
  options?: string[];
  correctOption?: number;
  timeLimit?: number;
  points?: number;
  orderIndex?: number;
}

export interface QuizWithQuestions {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  createdAt: string;
  updatedAt: string;
  questionCount: number;
  questions: Question[];
}

export type GameStatus = (typeof GameStatus)[keyof typeof GameStatus];

export const GameStatus = {
  waiting: "waiting",
  active: "active",
  finished: "finished",
} as const;

export interface Game {
  id: number;
  gameCode: string;
  quizId: number;
  quizTitle: string;
  status: GameStatus;
  /** @nullable */
  currentQuestionIndex: number | null;
  playerCount: number;
  createdAt: string;
}

export interface CreateGameBody {
  quizId: number;
}

export interface PlayerResult {
  playerId: number;
  nickname: string;
  score: number;
  rank: number;
  correctAnswers: number;
  totalAnswers: number;
}

export interface GameResults {
  gameCode: string;
  quizTitle: string;
  players: PlayerResult[];
}