# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuizBlast API - Kahoot Clone
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all quizzes
 */
export const ListQuizzesResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  questionCount: zod.number(),
});
export const ListQuizzesResponse = zod.array(ListQuizzesResponseItem);

/**
 * @summary Create a new quiz
 */
export const CreateQuizBody = zod.object({
  title: zod.string(),
  description: zod.string().nullish(),
});

/**
 * @summary Get a quiz with questions
 */
export const GetQuizParams = zod.object({
  id: zod.coerce.number(),
});

export const GetQuizResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  questionCount: zod.number(),
  questions: zod.array(
    zod.object({
      id: zod.number(),
      quizId: zod.number(),
      text: zod.string(),
      options: zod.array(zod.string()),
      correctOption: zod.number(),
      timeLimit: zod.number(),
      points: zod.number(),
      orderIndex: zod.number(),
    }),
  ),
});

/**
 * @summary Update a quiz
 */
export const UpdateQuizParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateQuizBody = zod.object({
  title: zod.string().optional(),
  description: zod.string().nullish(),
});

export const UpdateQuizResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  questionCount: zod.number(),
});

/**
 * @summary Delete a quiz
 */
export const DeleteQuizParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Add a question to a quiz
 */
export const AddQuestionParams = zod.object({
  id: zod.coerce.number(),
});

export const AddQuestionBody = zod.object({
  text: zod.string(),
  options: zod.array(zod.string()),
  correctOption: zod.number(),
  timeLimit: zod.number().optional(),
  points: zod.number().optional(),
  orderIndex: zod.number().optional(),
});

/**
 * @summary Update a question
 */
export const UpdateQuestionParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateQuestionBody = zod.object({
  text: zod.string().optional(),
  options: zod.array(zod.string()).optional(),
  correctOption: zod.number().optional(),
  timeLimit: zod.number().optional(),
  points: zod.number().optional(),
  orderIndex: zod.number().optional(),
});

export const UpdateQuestionResponse = zod.object({
  id: zod.number(),
  quizId: zod.number(),
  text: zod.string(),
  options: zod.array(zod.string()),
  correctOption: zod.number(),
  timeLimit: zod.number(),
  points: zod.number(),
  orderIndex: zod.number(),
});

/**
 * @summary Delete a question
 */
export const DeleteQuestionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Create a new game session from a quiz
 */
export const CreateGameBody = zod.object({
  quizId: zod.number(),
});

/**
 * @summary Get a game by code
 */
export const GetGameParams = zod.object({
  gameCode: zod.coerce.string(),
});

export const GetGameResponse = zod.object({
  id: zod.number(),
  gameCode: zod.string(),
  quizId: zod.number(),
  quizTitle: zod.string(),
  status: zod.enum(["waiting", "active", "finished"]),
  currentQuestionIndex: zod.number().nullable(),
  playerCount: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get final results for a completed game
 */
export const GetGameResultsParams = zod.object({
  gameCode: zod.coerce.string(),
});

export const GetGameResultsResponse = zod.object({
  gameCode: zod.string(),
  quizTitle: zod.string(),
  players: zod.array(
    zod.object({
      playerId: zod.number(),
      nickname: zod.string(),
      score: zod.number(),
      rank: zod.number(),
      correctAnswers: zod.number(),
      totalAnswers: zod.number(),
    }),
  ),
});