# lib/api-zod/src/generated/types/createQuestionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuizBlast API - Kahoot Clone
 * OpenAPI spec version: 0.1.0
 */

export interface CreateQuestionBody {
  text: string;
  options: string[];
  correctOption: number;
  timeLimit?: number;
  points?: number;
  orderIndex?: number;
}