# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuizBlast API - Kahoot Clone
 * OpenAPI spec version: 0.1.0
 */
import type { GameStatus } from "./gameStatus";

export interface Game {
  id: number;
  gameCode: string;
  quizId: number;
  quizTitle: string;
  status: GameStatus;
  /** @nullable */
  currentQuestionIndex: number | null;
  playerCount: number;
  createdAt: Date;
}