# lib/api-zod/src/generated/types/gameResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuizBlast API - Kahoot Clone
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerResult } from "./playerResult";

export interface GameResults {
  gameCode: string;
  quizTitle: string;
  players: PlayerResult[];
}