# lib/api-zod/src/generated/types/playerResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuizBlast API - Kahoot Clone
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerResult {
  playerId: number;
  nickname: string;
  score: number;
  rank: number;
  correctAnswers: number;
  totalAnswers: number;
}