# lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuizBlast API - Kahoot Clone
 * OpenAPI spec version: 0.1.0
 */

export interface Quiz {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  createdAt: Date;
  updatedAt: Date;
  questionCount: number;
}